Client/requirements.txt
PyQt5_sip == 12.9.0
